move tuning to regression methods
